# docs/api/api-spec.yaml
openapi: 3.0.3
info:
  title: Short URL API
  version: 1.1.0
  description: |
    统一返回结构为 ApiResponse<T>。
    成功业务码固定为 200，分页数据统一使用 PageResult 结构。

servers:
  - url: http://localhost:8080
    description: Local

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    ApiResponse:
      type: object
      required: [code, message, timestamp, requestId]
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: 操作成功
        data:
          nullable: true
        timestamp:
          type: string
          format: date-time
        requestId:
          type: string
          example: req_a1b2c3d4

    PageResult:
      type: object
      required: [list, total, page, size, totalPages]
      properties:
        list:
          type: array
          items:
            type: object
        total:
          type: integer
          format: int64
          example: 100
        page:
          type: integer
          example: 1
        size:
          type: integer
          example: 20
        totalPages:
          type: integer
          example: 5

paths:
  /api/shorten:
    post:
      summary: 创建短链
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [originalUrl]
              properties:
                originalUrl:
                  type: string
                title:
                  type: string
                expiredTime:
                  type: string
                  format: date-time
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

  /api/stats/{shortKey}:
    get:
      summary: 查询短链公开统计
      parameters:
        - name: shortKey
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

  /{shortKey}:
    get:
      summary: 短链重定向
      parameters:
        - name: shortKey
          in: path
          required: true
          schema:
            type: string
      responses:
        '302':
          description: redirect

  /api/auth/login:
    post:
      summary: 登录
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

  /api/auth/register:
    post:
      summary: 注册
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                password:
                  type: string
                email:
                  type: string
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

  /api/auth/check-username:
    get:
      summary: 检查用户名是否存在
      parameters:
        - name: username
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

  /api/auth/refresh-token:
    post:
      summary: 刷新令牌
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

  /api/auth/logout:
    post:
      summary: 退出登录
      security:
        - bearerAuth: []
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

  /api/auth/validate-token:
    get:
      summary: 校验令牌
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

  /api/auth/current-user:
    get:
      summary: 当前用户
      security:
        - bearerAuth: []
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

  /api/admin/users:
    get:
      summary: 用户分页列表
      security:
        - bearerAuth: []
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
    post:
      summary: 创建用户
      security:
        - bearerAuth: []
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

  /api/admin/users/{userId}:
    get:
      summary: 用户详情
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
    put:
      summary: 更新用户
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
    delete:
      summary: 删除用户
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

  /api/admin/users/{userId}/status:
    put:
      summary: 更新用户状态
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

  /api/admin/users/{userId}/reset-password:
    post:
      summary: 重置用户密码
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

  /api/admin/system-stats:
    get:
      summary: 系统统计
      security:
        - bearerAuth: []
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

  /api/admin/urls:
    get:
      summary: 短链分页列表
      security:
        - bearerAuth: []
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

  /api/admin/urls/{id}:
    get:
      summary: 短链详情
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
    put:
      summary: 更新短链
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
    delete:
      summary: 删除短链
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

  /api/admin/urls/{id}/status:
    put:
      summary: 更新短链状态
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

  /api/admin/urls/batch-status:
    put:
      summary: 批量更新短链状态
      security:
        - bearerAuth: []
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

  /api/admin/urls/batch:
    delete:
      summary: 批量删除短链
      security:
        - bearerAuth: []
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

  /api/admin/urls/{shortKey}/stats:
    get:
      summary: 管理端短链统计
      security:
        - bearerAuth: []
      parameters:
        - name: shortKey
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

  /api/admin/operation-logs:
    get:
      summary: 操作日志分页列表
      security:
        - bearerAuth: []
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

  /api/admin/operation-logs/stats:
    get:
      summary: 操作日志统计
      security:
        - bearerAuth: []
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

  /api/admin/dashboard/overview:
    get:
      summary: 仪表盘概览
      security:
        - bearerAuth: []
      parameters:
        - name: days
          in: query
          schema:
            type: integer
            default: 7
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

  /api/client-errors:
    post:
      summary: 前端错误日志上报
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [type, message]
              properties:
                type:
                  type: string
                message:
                  type: string
                code:
                  type: string
                status:
                  type: integer
                url:
                  type: string
                method:
                  type: string
                timestamp:
                  type: string
                  format: date-time
                details:
                  type: object
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

  /health:
    get:
      summary: 健康检查
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

  /health/detailed:
    get:
      summary: 详细健康检查
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

  /health/liveness:
    get:
      summary: 存活探针
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

  /health/readiness:
    get:
      summary: 就绪探针
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'